On parse failure, keep the old AST (#66)
Also, add a new `linesChangedSinceAST` attribute to the cache. This new attribute contains all lines of a document which have changed since the AST was last parsed
This allows us to still use the AST when the file is changed (a WIP is a common use-case)
In order to reduce the amount of bad definitions in that case, go-to-definition is only active if the currently targeted line hasn't changed. Otherwise, it's all messed up

While this feature is slightly useful for go-to-definition, it'll be absolutely necessary for the completion feature I'm working on